chore: migrate min-hash tests
